# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meta Earth Wallet Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Wallet {
  id: number;
  label: string;
  address: string;
  network: string;
  verified?: boolean;
  importSource?: string;
  createdAt: string;
}

export interface WalletInput {
  /** @minLength 1 */
  label: string;
  /** @minLength 1 */
  mnemonic?: string;
  /** @minLength 1 */
  privateKey?: string;
  /** @minLength 1 */
  password: string;
  network: string;
}

export interface WalletBalance {
  walletId: number;
  address: string;
  balance: string;
  denom: string;
  /** @nullable */
  usdValue?: string | null;
}

export interface AgentRunInput {
  walletIds: number[];
  masterPassword: string;
  dryRun?: boolean;
}

export interface AgentLog {
  id: number;
  /** @nullable */
  walletId?: number | null;
  /** @nullable */
  walletLabel?: string | null;
  action: string;
  status: string;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  amount?: string | null;
  message: string;
  createdAt: string;
}

export interface AgentRunResult {
  executed: number;
  skipped: number;
  logs: AgentLog[];
}

export interface AgentChatInput {
  message: string;
  /** @nullable */
  walletContext?: boolean | null;
}

export interface AgentChatResult {
  reply: string;
  suggestedActions?: string[];
}

export interface AgentStats {
  totalWallets: number;
  verifiedWallets?: number;
  totalWithdrawals: number;
  successfulWithdrawals: number;
  failedWithdrawals: number;
  activeRules: number;
  /** @nullable */
  lastRunAt?: string | null;
}

export type WalletStakingRewardsRewardsItem = {
  validatorAddress: string;
  amount: string;
  amountRaw: string;
};

export interface WalletStakingRewards {
  walletId: number;
  address: string;
  rewards: WalletStakingRewardsRewardsItem[];
  totalMEC: string;
  totalUMec: string;
}

export interface BulkImportInput {
  /** One mnemonic phrase per line (12 or 24 words each) */
  mnemonics: string;
  /** @minLength 1 */
  password: string;
  network?: string;
}

export type BulkImportResultWalletsVerifiedItem = {
  address?: string;
  label?: string;
  walletId?: number;
};

export type BulkImportResultWalletsUnverifiedItem = {
  address?: string;
  label?: string;
  walletId?: number;
};

export type BulkImportResultWallets = {
  verified?: BulkImportResultWalletsVerifiedItem[];
  unverified?: BulkImportResultWalletsUnverifiedItem[];
};

export type BulkImportResultSkippedDetailsItem = {
  reason?: string;
  phrase?: string;
};

export interface BulkImportResult {
  total: number;
  verified: number;
  unverified: number;
  skipped: number;
  wallets?: BulkImportResultWallets;
  skippedDetails?: BulkImportResultSkippedDetailsItem[];
}

export interface SweepConfig {
  id: number;
  masterAddress: string;
  enabled: boolean;
  autoClaimStaking: boolean;
  dividendWindowDays: number;
  minSweepAmountMec: string;
  updatedAt?: string;
}

export interface SweepConfigInput {
  masterAddress?: string;
  enabled?: boolean;
  autoClaimStaking?: boolean;
  dividendWindowDays?: number;
  minSweepAmountMec?: string;
}

export interface Rule {
  id: number;
  name: string;
  ruleType: string;
  enabled: boolean;
  /** @nullable */
  conditionJson?: string | null;
  /** @nullable */
  actionJson?: string | null;
  createdAt: string;
}

export interface RuleInput {
  /** @minLength 1 */
  name: string;
  ruleType: string;
  enabled?: boolean;
  conditionJson?: string;
  actionJson?: string;
}

/**
 * @nullable
 */
export type SchedulerStatusLastRunResult = {
  executed?: number;
  skipped?: number;
} | null;

export interface SchedulerStatus {
  enabled: boolean;
  /** @nullable */
  intervalMs?: number | null;
  walletIds: number[];
  dryRun: boolean;
  /** @nullable */
  nextRunAt?: string | null;
  /** @nullable */
  lastRunAt?: string | null;
  /** @nullable */
  lastRunResult?: SchedulerStatusLastRunResult;
}

export interface SchedulerInput {
  /** @minimum 60000 */
  intervalMs: number;
  walletIds: number[];
  masterPassword: string;
  dryRun?: boolean;
}

export interface RuleUpdate {
  name?: string;
  enabled?: boolean;
  conditionJson?: string;
  actionJson?: string;
}

export type WalletTransactionDirection = typeof WalletTransactionDirection[keyof typeof WalletTransactionDirection];


export const WalletTransactionDirection = {
  sent: 'sent',
  received: 'received',
} as const;

export interface WalletTransaction {
  txHash: string;
  height: number;
  timestamp: string;
  direction: WalletTransactionDirection;
  amount: string;
  amountRaw: string;
  counterpart: string;
  memo: string;
  success: boolean;
}

export interface WalletTransactionList {
  walletId: number;
  address: string;
  transactions: WalletTransaction[];
}

export interface WhitelistEntry {
  id: number;
  address: string;
  label: string;
  createdAt: string;
}

export interface WhitelistInput {
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  label: string;
}

export type GetWalletTransactionsParams = {
limit?: number;
};

export type ListAgentLogsParams = {
/**
 * @nullable
 */
walletId?: number | null;
/**
 * @nullable
 */
limit?: number | null;
};