# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meta Earth Wallet Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agent/sweep-config:
    get:
      operationId: getSweepConfig
      tags: [agent]
      summary: Get the current sweep configuration (master address, auto-sweep settings)
      responses:
        "200":
          description: Sweep config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SweepConfig"
    post:
      operationId: updateSweepConfig
      tags: [agent]
      summary: Update the sweep configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SweepConfigInput"
      responses:
        "200":
          description: Updated sweep config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SweepConfig"
  /agent/run:
    post:
      operationId: runAgent
      tags: [agent]
      summary: Trigger the AI agent to evaluate and execute actions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRunInput"
      responses:
        "200":
          description: Agent run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunResult"
  /agent/chat:
    post:
      operationId: agentChat
      tags: [agent]
      summary: Chat with the NVIDIA-powered agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentChatInput"
      responses:
        "200":
          description: Agent response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentChatResult"
  /agent/scheduler:
    get:
      operationId: getScheduler
      tags: [agent]
      summary: Get scheduler status
      responses:
        "200":
          description: Scheduler status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulerStatus"
    post:
      operationId: startScheduler
      tags: [agent]
      summary: Start or reconfigure the agent scheduler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchedulerInput"
      responses:
        "200":
          description: Scheduler started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulerStatus"
        "400":
          description: Invalid input
    delete:
      operationId: stopScheduler
      tags: [agent]
      summary: Stop the agent scheduler
      responses:
        "200":
          description: Scheduler stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulerStatus"
  /agent/logs:
    get:
      operationId: listAgentLogs
      tags: [agent]
      summary: Get agent activity logs
      parameters:
        - name: walletId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Agent logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentLog"
...
components:
  schemas:
...
    AgentRunInput:
      type: object
      required: [walletIds, masterPassword]
      properties:
        walletIds:
          type: array
          items:
            type: integer
        masterPassword:
          type: string
        dryRun:
          type: boolean
          default: false
    AgentRunResult:
      type: object
      required: [executed, skipped, logs]
      properties:
        executed:
          type: integer
        skipped:
          type: integer
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AgentLog"
    AgentChatInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        walletContext:
          type: ["boolean", "null"]
    AgentChatResult:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
        suggestedActions:
          type: array
          items:
            type: string
    AgentLog:
      type: object
      required: [id, action, status, message, createdAt]
      properties:
        id:
          type: integer
        walletId:
          type: ["integer", "null"]
        walletLabel:
          type: ["string", "null"]
        action:
          type: string
        status:
          type: string
        txHash:
          type: ["string", "null"]
        amount:
          type: ["string", "null"]
        message:
          type: string
        createdAt:
          type: string
          format: date-time
...
    SweepConfig:
      type: object
      required: [id, masterAddress, enabled, autoClaimStaking, dividendWindowDays, minSweepAmountMec]
      properties:
        id:
          type: integer
        masterAddress:
          type: string
        enabled:
          type: boolean
        autoClaimStaking:
          type: boolean
        dividendWindowDays:
          type: integer
        minSweepAmountMec:
          type: string
        updatedAt:
          type: string
          format: date-time
    SweepConfigInput:
      type: object
      properties:
        masterAddress:
          type: string
        enabled:
          type: boolean
        autoClaimStaking:
          type: boolean
        dividendWindowDays:
          type: integer
        minSweepAmountMec:
          type: string
...
    SchedulerStatus:
      type: object
      required: [enabled, walletIds, dryRun]
      properties:
        enabled:
          type: boolean
        intervalMs:
          type: ["integer", "null"]
        walletIds:
          type: array
          items:
            type: integer
        dryRun:
          type: boolean
        nextRunAt:
          type: ["string", "null"]
        lastRunAt:
          type: ["string", "null"]
        lastRunResult:
          type: ["object", "null"]
          properties:
            executed:
              type: integer
            skipped:
              type: integer
    SchedulerInput:
      type: object
      required: [intervalMs, walletIds, masterPassword]
      properties:
        intervalMs:
          type: integer
          minimum: 60000
        walletIds:
          type: array
          items:
            type: integer
        masterPassword:
          type: string
        dryRun:
          type: boolean
          default: true
...