# lib/api-zod/src/generated/types/sweepConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meta Earth Wallet Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface SweepConfig {
  id: number;
  masterAddress: string;
  enabled: boolean;
  autoClaimStaking: boolean;
  dividendWindowDays: number;
  minSweepAmountMec: string;
  updatedAt?: Date;
}